+++ pkgs/dart_tooling_mcp_server/lib/src/server.dart
@@ -16,6 +16,7 @@ import 'mixins/dash_cli.dart';
 import 'mixins/dtd.dart';
 import 'mixins/pub.dart';
 import 'mixins/pub_dev_search.dart';
+import 'mixins/roots_fallback_support.dart';
 import 'utils/file_system.dart';
 import 'utils/process_manager.dart';
 
@@ -26,6 +27,7 @@ final class DartToolingMCPServer extends MCPServer
         ToolsSupport,
         ResourcesSupport,
         RootsTrackingSupport,
+        RootsFallbackSupport,
         DartAnalyzerSupport,
         DashCliSupport,
         PubSupport,
@@ -36,6 +38,7 @@ final class DartToolingMCPServer extends MCPServer
     super.channel, {
     @visibleForTesting this.processManager = const LocalProcessManager(),
     @visibleForTesting this.fileSystem = const LocalFileSystem(),
+    this.forceRootsFallback = false,
   }) : super.fromStreamChannel(
          implementation: ServerImplementation(
            name: 'dart and flutter tooling',
@@ -47,9 +50,13 @@ final class DartToolingMCPServer extends MCPServer
        );
 
   static Future<DartToolingMCPServer> connect(
-    StreamChannel<String> mcpChannel,
-  ) async {
-    return DartToolingMCPServer(mcpChannel);
+    StreamChannel<String> mcpChannel, {
+    bool forceRootsFallback = false,
+  }) async {
+    return DartToolingMCPServer(
+      mcpChannel,
+      forceRootsFallback: forceRootsFallback,
+    );
   }
 
   @override
@@ -57,4 +64,7 @@ final class DartToolingMCPServer extends MCPServer
 
   @override
   final FileSystem fileSystem;
+
+  @override
+  final bool forceRootsFallback;
 }