+++ pkgs/dart_tooling_mcp_server/lib/src/utils/constants.dart
@@ -2,11 +2,14 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:dart_mcp/server.dart';
+
 /// A namespace for all the parameter names.
 extension ParameterNames on Never {
   static const column = 'column';
   static const command = 'command';
   static const line = 'line';
+  static const name = 'name';
   static const packageName = 'packageName';
   static const paths = 'paths';
   static const position = 'position';
@@ -14,4 +17,8 @@ extension ParameterNames on Never {
   static const root = 'root';
   static const roots = 'roots';
   static const uri = 'uri';
+  static const uris = 'uris';
 }
+
+/// A shared success response for tools.
+final success = CallToolResult(content: [Content.text(text: 'Success')]);